Initial version of tc network filter for host isolation
Add initial version of tc (traffic control) eBPF program.

The program starts filtering the network as soon as it is loaded.
All egress TCP packets are blocked unless destination IP is in the 'allowed_IPs' map.
All egress UDP packets are blocked except for DNS port.
ARP packets are allowed. All other type of traffic is blocked.

No filtering is done on the receive path.

The program should be attached to tc in the egress path.

Installed kernel headers are used by default. There is a macro switch in
kerneldefs.h which allows using only static definitions in the header
without requiring kernel headers to be installed on the machine.