Switch to BTF-style maps
Legacy eBPF map definitions are planned to be removed in libbpf v1.0. Switch to BTF maps in the tc filter.

Note that BTF maps can be used without BTF being enabled in the kernel. BTF-defined maps are only handled by libbpf. It is a feature which allows annotating maps with type info for more human-readable dumps in bpftool for example.
However, BTF maps add some constraints - they require newer Clang (7+) and you need to compile with debug symbols (-g) as well, in order for the .BTF sections to be created.
More constraints: tc has to be compiled with libbpf, otherwise it fails to relocate the maps.
When using new Clang, the flag -mcpu=v2 needs to be used to prevent it from generating 32-bit instructions (which fail to load on older kernels).

This change was tested on Ubuntu 16 and 20 with Clang-12. Same as before, the eBPF binary can be moved across systems and it loads successfully.