Switch to static headers
Use local static headers instead of relying on linux/* kernel headers being installed.

The local kerneldefs.h header contains only the necessary definitions which TcFilter needs.
The header should be portable (at least for x64) as it mostly contains network protocol definitions which are not going to change in the kernel.

Add missing dependency on kerneldefs.h in the Makefile.