Make Razor run eBPF build commands every time
CustomCommands used for building eBPF subprojects will be run every time regardless of dependencies.
This is ok because the CustomCommands mostly run 'make' which won't do anything if source files haven't changed.
One exception is elftoolchain which has a bug which makes it rebuild every time bmake is run.
However, elftoolchain sources won't ever be modified (there is only one .patch file for the Makefile) so we will skip building it if the build directory already exists.