Revert "Do not drop all fragmented packets"
This reverts commit 1b68e6568ac8cedcae416c1f5e86e7523599b1b8.

The fragmentation check needs to remain for the time being because
otherwise the L4 parsing code will be incorrect (it assumes the
TCP/UDP header exists if IP protocol field is set to TCP/UDP,
which is not true for non-first fragments).

TODO later: in order to support fragmented packets, we must skip the TCP/UDP
parsing code for them and allow/reject them based only on IP allowlist.