Add LPM_TRIE map containing allowed subnets
This map will contain user-defined allowed subnets.
TcFilter first checks if packet destination IP is in any allowed subnet.
If not, it searches the allowed_IPs map with single allowed IPs.

The LPM_TRIE map type simplifies the lookup even though it may not be the most optimal
(we do not care about the Longest Prefix Match, just any prefix match), but an array
or hash map would be cumbersome to use because we cannot have loops.

LPM_TRIE is not preallocated so it does not take up space if empty.

This map type is available since kernel 4.11.