tests: BPFTcFilterTests map unpin on teardown
Since the maps are pinned to the filesystem
the tests might be flaky if the maps are not cleared at every iteration.

This test could be made more pure by non pinning the map. However we
really want to test that behavior as well because we rely on pinning and
non pinned maps could behave differently because they have a different
lifecycle.