Add "libebpf" header implementation
This commit adds a basic implementation to go along with the file found in
include/libebpf.h.

The library was renamed to "LibEbpfEvents" (now found under
non-GPL/LibEbpfEvents) to be more explicit about what it does. it
currently uses a single set of ringbuffer probes regardless of the
features/events requested and only supports the BPF ringbuffer.

The actual wire protocol has been factored out of LibEbpfEvents and into
a dual-licensed header at GPL/Events/EbpfEventProto.h. This makes for
easy including in the probes (no janky #ifdefs to exclude userspace
code when including it in the probes) and easy dual-licensing of only
the part of the events system that needs to be dual-licensed.

This commit also adds very basic process fork and exec events to the
existing events probe and renames the FileEvents probe to simply Probe.
We can split the probes off later, but for this iteration, the simplest
thing to do seems to be just to have all event types in a single probe.