new(GPL/EventProbe): implement EBPF_EVENT_FILE_DELETE
This is an alternative implementation to the `security_path_unlink` one
we were exploring until now. That implementation needs `CONFIG_SECURITY`
to work which is not present on most systems yet.

This also includes a mechanism to save the state of the current
file event between multiple programs using the tid.

This still has some limitations:
- we need to instruct the dentry parameter for vfs unlink from userspace
  or via btf checks
- there are some bugs with vfs mount (some paths are duplicated)