Fix paths passed to userspace containing garbage
The buf argument to the dentry resolver points into a ringbuffer, and
is thus uninitialized. If we don't zero it out, the buf[0] == '\0' check
at the end of the function might see a garbage, non-null byte in buf[0]
and assume a path has been constructed, leaving it as garbage. We need
to zero out the buffer.