Clean up Repository for Better Presentation to the General Public (#113)
* Unify userspace hostisolation code under one dir

* Unify userspace events code under one dir

* Make GPL/ dir structure the same as non-GPL/

Such that we have GPL/Events and non-GPL/Events (and the same for
HostIsolation). Delegate more work to subdirectory CMakeLists.txt files
to clean up the root CMakeLists.txt.

* Remove docs/

This was only holding one feature proposal and more aren't likely to be
added since that seems to be being done outside of VCS now. We'll still
have this doc in git history so just remove the dir to cleanup the
directory structure.

* Rename LibEbpfEvents to EbpfEvents

CMake will automatically prepend "Lib" to the name of something created
with add_library, resulting in a artifact called "LibLibEbpfEvents",
this has been bugging me for some time.

Also rename "LibEbpfEvents" directory to simply "Lib" for consistency
with HostIsolation directory, which puts all its userspace components in
a "Lib" directory.

* Remove GPL/Events/BPFEventsTests.cpp

This file was never used and is unlikely to be used now that we've got a
pretty comprehensive multikernel testing setup. If needed, it can be
re-checked out from CI, no use in keeping dead code around.

* Remove unneeded kernel re-definitions

A relic from the days when building this repo was tied to the endpoint
toolchain and it's ancient kernel headers required this kludge.

* Remove stale TODOs

Endpoint is currently running fine with a 256kb ringbuffer. This can be
changed in the future, but for now isn't causing any problems.

* Update README.md, move more specific docs to docs/

* Fix incorrect SPDX license identifiers

Both LicenseRef-Elastic-License-2.0 and Elastic-License-2.0 are
incorrect. The correct format is Elastic-2.0.

* Don't check code formatting in Jenkins

This is done in GitHub actions now.

* Fix EventsTrace path for new dir structure

* Fix typo in README.md

Co-authored-by: Nicholas Berlin <56366649+nicholasberlin@users.noreply.github.com>

* Fix typo in README.md

Co-authored-by: Nicholas Berlin <56366649+nicholasberlin@users.noreply.github.com>

* Remove TARGET_TMP_DIR (unused)

* s/BPF/eBPF/g

Co-Authored-By: Mattia Meleleo: <mattia.meleleo@elastic.co>

* Fix incorrect kernel support version

Co-Authored-By: Mattia Meleleo <mattia.meleleo@elastic.co>

* Remove test loader tool

Unneeded now that we have a full multi kernel test setup.

Co-Authored-By: Mattia Meleleo <mattia.meleleo@elastic.co>

* Clarify vmlinux docs

Using the "prebuilt" vmlinux.h is the default now. This shouldn't affect
functionality in any way. Update the docs to reflect these two facts.

Co-authored-by: Nicholas Berlin <56366649+nicholasberlin@users.noreply.github.com>
Co-authored-by: Mattia Meleleo: <mattia.meleleo@elastic.co>