Fix endpoint build issues (#116)
* Remove -lz from library link targets

With this being the case, anything that links to libbpf will
*dynamically* link libz, which we don't want for endpoint. Remove -lz
from libbpf and manually add it to all executables that need it, with
a comment to remove later when we vendor in libz.

* ifdef-out tty_write probe

The tty write probe (specifically the function-prototype detection
logic) is suspect to be causing test failures in endpoint. Ifdef it out
for the time being so we have a working build. Once the build is
confirmed to be fixed we can fix the BTF logic.

* Fix docker issues

- Push new container as current one was overwritten with something weird
- Fix broken container and tag-container targets
- Run clang-format in working container