Add error-check for BTF load call
Our call to btf__load_vmlinux_btf was not being error checked with
libbpf_error(...) and returning a bad pointer on systems with no BTF,
which was then used, and caused a segfault.

Also add libEbpfEvents-level logging function that logs to stderr
if verbose libbpf logging is turned on. This will make stuff like this
much easier to catch in CI.