Fix probes not loading on PREEMPT_RT kernels
On kernels with the PREEMPT_RT patch, the trace_entry struct, which
is the first member of trace_event_raw_sys_exit struct (the ctx for the
setsid exit probe) has two extra members.

We access the ret field of ctx via ctx->ret, which compiles to a
hardcoded offset. We should be using BPF_CORE_READ such that a
relocation is emitted so the correct offset is accessed on rt kernels.