Run host isolation tests in GitHub Actions
Completely break free of Jenkins by running host isolation tests in
GitHub actions. The existing jenkins setup was broken anyways and slated
to be moved to the multikernel tester.

As a bonus of moving the host isolation tests to the new test framework,
they're now running on far more kernels than before.