De-generify feature logic
Since it's existed, the "features" argument passed into
ebpf_event_ctx__create via the opts struct has been used for the sole
purpose of bpf trampoline detection. There's also no scenario currently
where we have to force kprobes over trampolines (i.e. the library always
performs the detection logic), meaning having to go through the two-step
process of ebpf_detect_system_features and ebpf_event_ctx__new is an
unncecessary level of indirection.

There is one use case for which *knowing* (but not setting) the features
is useful, the TestFeaturesCorrect multi-kernel test case. To ensure we
can still test the feature detection logic, add
ebpf_event_ctx__get_features to expose the features detected by the
library, while removing the ability to set them.

If in the future, more features come along and it becomes necessary to
toggle them on a rather granular basis, we can revert this, but for the
time being, it simplifies the code nicely.