Move support detection logic to this repository
Since the very beginning, the logic to detect if we support BPF has been
done in endpoint instead of in the ebpf events library. This is wonky.
Logically, this logic belongs here, the library should be performing all
necessary checks and returning -ENOTSUP if it can't load the probes.
library users shouldn't have to divine the specific system
characteristics needed for the library to successfully run themselves.