Cut down support detection logic
This reduces the stuff we check for to just
bpf_support && (kernel_version > 5.10.16).

While this doesn't cover 100% of cases, it will likely cover 99.9% of
cases and give us a easy to read error message in those cases. Probe
loading will ultimately fail if we're e.g. on a 5.10.16+ kernel with BTF
that doesn't have kprobes or ftrace enabled, which will cause endpoint
to not use BPF.