Implement variable-length ringbuf wire format
This commit introduces variable-length fields into the ringbuf wire
format.

Over time, the number of fields in our wire format with a large
theoretical maximum size has grown (pathnames, argv, tty output, etc.).

Every field in the wire format is of a fixed-size. This simplifies
parsing of messages in userspace, but leads to huge memory waste as the
maximum possible buffer size needs to be allocated, even when it's not
used most of the time. For example, PATH_MAX on linux is 4096, but the
pathnames we pass up are going to be a fraction of that size 99% of the
time. Terminal output is an even bigger culprit for this sort of thing,
a 4 KiB buffer is allocated for each tty_write event, even if the data
being written to the tty is just one byte. For a high-frequency tty
write event, this can easily cause us to drop events.

This commit adds a few new structures present at the end of each event
struct that needs to use variable length fields. Fields that are always
of a fixed size (or variable length but so small as to be irrelevant
from a memory-usage perspective, e.g. task->comm) remain as they were
before, in a fixed-size part of the event. Variable length events are
packed into a flexible array member at the end of the event struct and
each are preceeded by a header describing the type of field and size.

An unfortunate necessity of this approach is that we cannot use the
ringbuffer reserve/commit API for events that have variable-length
fields, as we won't know the size of the event in advance. Instead, we
build each event with its variable-length fields in a percpu map, and
then output it using bpf_ringbuf_output. This should result in a small
performance hit as we're introducing an additional memory copy, but
should be made up for in massive memory usage improvements. An
alternative would have been to continue to use the reserve/commit API,
passing up each event in fixed-size chunks, but this would just shift
the extra copy to userspace, and introduce a bunch of complexity there
as we'd need to have reconstruction logic in the Ebpf events library. In
any case, there is no free lunch.

Logic and helpers for adding variable length parameters have been added
to Helpers.h and are fairly well documented. New events that need to use
variable-length fields should add a ebpf_varlen_fields_start struct as
their last member with a name of vl_fields. Variable-length fields can
then be added to the event. For example, in the case of a file delete
event:

        struct ebpf_file_delete_event *event = get_event_buffer();

        // ... Fill in fixed-length fields

        // Variable length fields
        ebpf_vl_fields__init(&event->vl_fields);
        struct ebpf_varlen_field *field;
        size_t size;

        // path
        ADD_VL_FIELD_OR_GOTO_EMIT(event, field, EBPF_VL_FIELD_PATH);
        size = ebpf_resolve_path_to_string(field->data, &p, task);
        ebpf_vl_field__set_size(&event->vl_fields, field, size);

    emit:
        bpf_ringbuf_output(&ringbuf, event, EVENT_SIZE(event), 0);

The emit label is required. The ADD_VL_FIELD_OR_GOTO_EMIT macro will
jump to it if there is no more space remaining in the event buffer
returned from get_event_buffer().

The use of variable-length fields also allows us to get rid of a bit of
cruft. In particular:

- No more PATH_MAX * 2 sized buffers in EbpfEventProto.h that waste
  space and leak an implementation detail of the BPF probes to userspace
- The memset reimplementation to deal with argv has been removed as we
  will now know size of argv in the variable-length parameter header,
  making zeroing out a fixed-size buffer unnecessary

This commit also bumps the ringbuffer size to 4 MiB so that it's larger
than the 2 MiB percpu event buffers.