Remove ADD_VL_PARAM_OR_GOTO_EMIT
Given the caps on argv and tty_out, the maximum theoretical possible
event size is currently ~40 KiB. Given 128 KiB of useable space in the
percpu event buffer, this means we will never run out of space and the
goto emit path will never be hit, making this logic unnecessary for the
time being. If in the future we need to read truly massive fields, this
can be revisited and potentially re-added.