Hook exit only after descriptors are closed. Fixes #216 (#218)
Previously we were sending exit events before the file descriptors were closed.
The tests are likely naive enough to always properly close the socket and then
terminate. This is easy to see by connecting and killing it.

The solution is to move the hook down to sched_process_exit(). In order to
respect the group_dead dance we have to now keep state of who got it, we use an
empty state, meaning the existence of the entry in the map is the state.