Fix TCP close events for "idle" connections. Fixes #217
Change the condition to supress a close, before it would supress any sockets
that had no traffic.

Now we keep state of sockets we sent an attempt/accept and supress the ones we
didn't track _AND_ had no traffic. The map is limited, once we have 128k active
sockets we can't track them, so fallback to looking at bytes_received and
bytes_sent.

We don't really look at tgid, but we will on upcoming probes.
Maps are not small, most of the allocation size goes to the hash bucket. The
empty map wires 2MB of memory.

6306: hash  name sk_to_tgid  flags 0x1
        key 8B  value 4B  max_entries 131072  memlock 2098760B
        btf_id 5117
        pids EventsTrace(2868637)